[leaving esk] Move away from Emacs Starter Kit
I removed starter-kit-lisp as a package, and this in turn stopped
bringing in starter-kit.

I think this is the direction to move in ala v3 of ESK: it's all prose
now.

So, if I no longer have starter-kit, then I need to bring in magit
explicitly, and I'm going to try out 'better-defaults'-- the one package
recommended by technomancy now.

Without magit things didn't start at all. Now, things start, but there
will cleary be a bunch of upcoming changes to "get back" what the
starter-kit used to bring me.